Crud role refactoring
